+++ library/src/schemas/objectWithRest/objectWithRest.test-d.ts
@@ -1,9 +1,15 @@
 import { describe, expectTypeOf, test } from 'vitest';
-import type { ReadonlyAction, TransformAction } from '../../actions/index.ts';
+import type {
+  Brand,
+  BrandAction,
+  ReadonlyAction,
+  TransformAction,
+} from '../../actions/index.ts';
 import type { SchemaWithPipe } from '../../methods/index.ts';
 import type { InferInput, InferIssue, InferOutput } from '../../types/index.ts';
 import type { AnySchema } from '../any/index.ts';
 import type { BooleanIssue, BooleanSchema } from '../boolean/index.ts';
+import type { CustomIssue, CustomSchema } from '../custom/index.ts';
 import type { ExactOptionalSchema } from '../exactOptional/index.ts';
 import type { NullishSchema } from '../nullish/index.ts';
 import {
@@ -68,6 +74,10 @@ describe('objectWithRest', () => {
             TransformAction<undefined | string, number>,
           ]
         >;
+        key07: CustomSchema<`a${string}` | `b${string}`, undefined>;
+        key08: SchemaWithPipe<
+          [StringSchema<undefined>, BrandAction<string, 'foo'>]
+        >;
 
         // ExactOptionalSchema
         key10: ExactOptionalSchema<StringSchema<undefined>, undefined>;
@@ -103,6 +113,8 @@ describe('objectWithRest', () => {
           key04: string;
           key05: string | undefined;
           key06?: string | undefined;
+          key07: `a${string}` | `b${string}`;
+          key08: string;
 
           // ExactOptionalSchema
           key10?: string;
@@ -137,6 +149,8 @@ describe('objectWithRest', () => {
           readonly key04: string;
           key05: string;
           key06?: number;
+          key07: `a${string}` | `b${string}`;
+          key08: string & Brand<'foo'>;
 
           // ExactOptionalSchema
           key10?: string;
@@ -167,6 +181,7 @@ describe('objectWithRest', () => {
         | StringIssue
         | NumberIssue
         | BooleanIssue
+        | CustomIssue
       >();
     });
   });