+++ packages/to-json-schema/src/toJsonSchema.test.ts
@@ -2,6 +2,8 @@ import * as v from 'valibot';
 import { describe, expect, test, vi } from 'vitest';
 import { toJsonSchema } from './toJsonSchema.ts';
 
+// TODO: Add tests for override configs
+
 console.warn = vi.fn();
 
 describe('toJsonSchema', () => {
@@ -44,6 +46,53 @@ describe('toJsonSchema', () => {
           },
         },
       });
+    });
+
+    test('for complex schema with any order of definitions', () => {
+      const stringSchema = v.string();
+      const aliasesSchema = v.array(stringSchema);
+      const complexSchema = v.pipe(
+        v.object({
+          name: v.lazy(() => stringSchema),
+          aliases: v.optional(aliasesSchema),
+          email: v.pipe(stringSchema, v.email(), v.minLength(10)),
+        }),
+        v.description('foo')
+      );
+      const expectedJsonSchema = {
+        $schema: 'http://json-schema.org/draft-07/schema#',
+        $ref: '#/$defs/complexSchema',
+        $defs: {
+          stringSchema: { type: 'string' },
+          aliasesSchema: {
+            type: 'array',
+            items: { $ref: '#/$defs/stringSchema' },
+          },
+          complexSchema: {
+            type: 'object',
+            properties: {
+              name: { $ref: '#/$defs/stringSchema' },
+              aliases: { $ref: '#/$defs/aliasesSchema' },
+              email: { type: 'string', format: 'email', minLength: 10 },
+            },
+            required: ['name', 'email'],
+            description: 'foo',
+          },
+        },
+      };
+      const definitionPermutations = [
+        { stringSchema, aliasesSchema, complexSchema },
+        { stringSchema, complexSchema, aliasesSchema },
+        { aliasesSchema, stringSchema, complexSchema },
+        { aliasesSchema, complexSchema, stringSchema },
+        { complexSchema, stringSchema, aliasesSchema },
+        { complexSchema, aliasesSchema, stringSchema },
+      ];
+      for (const definitions of definitionPermutations) {
+        expect(toJsonSchema(complexSchema, { definitions })).toStrictEqual(
+          expectedJsonSchema
+        );
+      }
     });
 
     test('for recursive schema', () => {